pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chatgpt-library-archiver"
version = "0.1.0"
description = "Toolset for downloading, archiving, and browsing images generated via ChatGPT."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
  {name = "Kyle Johnson", email = "me@kyle.engineer"}
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Topic :: Utilities"
]
keywords = [
  "chatgpt",
  "openai",
  "image",
  "archiver",
  "gallery",
  "download"
]
dependencies = [
  "requests>=2.31.0",
  "tqdm>=4.66.0",
  "openai>=1.0.0",
  "Pillow>=10.0.0",
  "cryptography>=41.0.0"
]

[project.optional-dependencies]
dev = [
  "build>=1.0.0",
  "pre-commit>=3.5.0",
  "pytest>=7.0",
  "pytest-cov>=4.1",
  "pyright>=1.1.390",
  "ruff>=0.5.0",
  "Pillow>=10.0.0"
]

[project.scripts]
chatgpt-archiver = "chatgpt_library_archiver.__main__:main"

[project.urls]
Homepage = "https://github.com/kabaka/chatgpt-library-archiver"
Repository = "https://github.com/kabaka/chatgpt-library-archiver"
Issues = "https://github.com/kabaka/chatgpt-library-archiver/issues"

[tool.pytest.ini_options]
addopts = "-q -m 'not slow'"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselected by default, run with -m slow)",
    "integration: marks integration tests requiring external resources",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "PL", "RUF", "FURB", "S", "PTH"]

[tool.ruff.lint.per-file-ignores]
"src/chatgpt_library_archiver/ai.py" = ["PLR0913"]
"src/chatgpt_library_archiver/cli/app.py" = ["PLR0913"]
"src/chatgpt_library_archiver/http_client.py" = ["PLR0913"]
"src/chatgpt_library_archiver/incremental_downloader.py" = ["PLC0415", "PLR0912", "PLR0913", "PLR0915"]
"src/chatgpt_library_archiver/tagger.py" = ["PLR0912", "PLR0913", "PLR0915"]
"src/chatgpt_library_archiver/thumbnails.py" = ["PLR0912", "PLR0915"]
"src/chatgpt_library_archiver/bootstrap.py" = ["S603", "S607"]
"src/chatgpt_library_archiver/browser_extract.py" = ["PLC0415", "PLR0913", "PTH101", "S603", "S607", "SLF001"]
"src/chatgpt_library_archiver/cli/commands/extract_auth.py" = ["PLC0415"]
"src/chatgpt_library_archiver/metadata.py" = ["PTH105", "PTH108"]
"src/chatgpt_library_archiver/utils.py" = ["PTH101"]

"tests/*.py" = ["PLC0415", "PLR2004", "PTH", "S101", "S105", "S603", "S607"]

[tool.ruff.lint.isort]
known-first-party = ["chatgpt_library_archiver"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.coverage.run]
source = ["chatgpt_library_archiver"]

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "warning"
venvPath = "."
venv = ".venv"
include = [
  "src/chatgpt_library_archiver/metadata.py",
  "src/chatgpt_library_archiver/status.py",
  "src/chatgpt_library_archiver/utils.py",
  "src/chatgpt_library_archiver/ai.py",
  "src/chatgpt_library_archiver/http_client.py",
  "src/chatgpt_library_archiver/gallery.py",
  "src/chatgpt_library_archiver/thumbnails.py",
]

[tool.setuptools.package-data]
chatgpt_library_archiver = ["gallery_index.html"]

[tool.setuptools]
include-package-data = true
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
